Created utils.py
