+++ requirements.txt
@@ -1,3 +1,5 @@
 opencv-python
 pyautogui
 mediapipe
+numpy
+pynput
